New unit test: addAndDelete

+++ src/folderEncryption.test.ts
@@ -1,6 +1,6 @@
 import test from 'node:test';
 import fs from 'fs';
-import assert from 'node:assert';
+import assert from 'node:assert/strict';
 import { compareSync } from 'dir-compare';
 import { FolderEncryption, FolderEncryptionStats } from './folderEncryption';
 
@@ -32,6 +32,36 @@ test(FolderEncryption.prototype.sync.name, function () {
 	assert.equal(comparison.same, true);
 	assert.equal(comparison.total, 5);
 
-	fs.rmSync('./test.1', { recursive: true });
-	fs.rmSync('./test.0', { recursive: true });
+	if (fs.existsSync('./test.1')) fs.rmSync('./test.1', { recursive: true });
+	if (fs.existsSync('./test.0')) fs.rmSync('./test.0', { recursive: true });
+});
+
+test(FolderEncryption.prototype.sync.name + '.addAndDelete', function () {
+	if (fs.existsSync('./test.1')) fs.rmSync('./test.1', { recursive: true });
+	if (fs.existsSync('./test.0')) fs.rmSync('./test.0', { recursive: true });
+
+	let folderEncryption = new FolderEncryption('password', './test', './test.1');
+	folderEncryption.sync();
+
+	fs.writeFileSync('./test/new.txt', 'test');
+	folderEncryption.sync();
+	assert.equal(folderEncryption.stats.newFiles, 1);
+	assert.equal(folderEncryption.stats.deletedFiles, 0);
+	new FolderEncryption('password', './test.1', './test.0').unpack();
+	let comparison = compareSync('./test', './test.0', { compareContent: true });
+	assert.equal(comparison.same, true);
+	assert.equal(comparison.total, 6);
+
+	fs.unlinkSync('./test/new.txt');
+	folderEncryption.sync();
+	assert.equal(folderEncryption.stats.newFiles, 0);
+	assert.equal(folderEncryption.stats.deletedFiles, 1);
+	if (fs.existsSync('./test.0')) fs.rmSync('./test.0', { recursive: true });
+	new FolderEncryption('password', './test.1', './test.0').unpack();
+	comparison = compareSync('./test', './test.0', { compareContent: true });
+	assert.equal(comparison.same, true);
+	assert.equal(comparison.total, 5);
+
+	if (fs.existsSync('./test.1')) fs.rmSync('./test.1', { recursive: true });
+	if (fs.existsSync('./test.0')) fs.rmSync('./test.0', { recursive: true });
 });
